alterando tabelas, apagando tabelas desnecessarias, criando nova fato

+++ dbt_dw_ans/models/trusted/tr_caracteristicas_produtos_suplementares.sql
@@ -27,5 +27,5 @@ FROM {{source('stage', 'caracteristicas_produtos_suplementares') }} cps
 WHERE EXISTS (
     SELECT 1 
     FROM {{ ref('tr_areas_comercializacao_planos') }} acp
-    WHERE acp.id_plano = cps.id_plano
+    WHERE acp.id_plano = cps.id_plano::INTEGER
 )
